Write first failing test to verify file setup
I did this because I wanted to make sure the requires that I am using in this
directory are correct. When I run `ruby
bracer_validator/bracer_validator_test.rb` the test fails becasue the validate
method does not exist yet. I will start working on this and try to commit when
it seems appropriate.