Start building command method for validate
I decided that I would create a command method to do the calculations for
discoving if a given string is valid. I want the `valid?` method to access
that data and return true or false based on `validate`s findings. I am
starting to try and test behavior and decided to use a mock. This is proving
to be harder than I expected. I had to be sure to inject the mock object into
my code so that I could verify that the chars method is being called.